Use overlapped IO for the ConPTY signal pipe.
Replace CreatePipe with CreateOverlappedPipe and adapt signal reads/writes so the pipe ends support asynchronous IO. Fixes #20528.
