+++ src/types/lib/types.vcxproj
@@ -28,6 +28,7 @@
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="..\utils.cpp" />
+    <ClCompile Include="..\overlappedPipe.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\IControlAccessibilityInfo.h" />
