+++ src/winconpty/ft_pty/winconpty.FeatureTests.vcxproj
@@ -23,6 +23,9 @@
     <ProjectReference Include="..\lib\winconptylib.vcxproj">
       <Project>{58a03bb2-df5a-4b66-91a0-7ef3ba01269a}</Project>
     </ProjectReference>
+    <ProjectReference Include="..\..\types\lib\types.vcxproj">
+      <Project>{18D09A24-8240-42D6-8CB6-236EEE820263}</Project>
+    </ProjectReference>
   </ItemGroup>
   <ItemDefinitionGroup>
     <ClCompile>
